+++ documentation/Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## General Information
+
+Some versions in the changelog were skipped because they were released or pushed 
+as part of a later version.
+Time Copy was only converted into a side-bar version from version 0.8.7.
+
 ## V 0.8.80
 - fix: text-label styling on small resolution
 - feature: get new "Master-Booking-Number" marked with "°" as Ticketnumber, when exists.
@@ -122,3 +128,18 @@
 - other: renaming low latency to high latency
 - other: new public file structure
 - other: readme, license, documentation
+
+## V 1.0.0
+- feature: show / hide messages
+- feature: amagprotime dlc - multiple string matches in detection item
+- feature: compatibility with older profile versions / supported profile version list
+- feature: return message, if browser is offline
+- fix: show error notification when trying to import currupt profile
+- fix: show error notification after reimport same profile
+- fix: remove automatic error json feedback
+- fix: write default values into exported file on fresh profile
+- fix: font fallback
+- other: new profile version
+- other: new Theme "Japan's White"
+- other: documentation
+- other: restructure dlc cache